fix(admin-db): Phase-1 schema hardening migration [findings 010-019, 023-033]
Restores CHECK constraints, partial indexes, partial unique indexes, FK
relationships, and missing columns/indexes that the C# port shed during
the TS → EF migration. New EF migration `20260419015726_SchemaHardeningPhase1`
mixes EF model changes with raw `migrationBuilder.Sql` for Postgres-specific
features (partial indexes, CHECK, partial unique on nullable column).

Per-finding outcomes:

- 010: Invalid (intentional simplification — user_installations folded into tenant_memberships)
- 011: Fixed (DROP-list now includes user_api_keys, user_installations, tenant_invites, email_outbox, queued_tasks)
- 012: Fixed (refresh_tokens active_expires partial index)
- 013: Fixed (password_reset_tokens active_expires partial index)
- 014: Fixed (sanitization_rules: partial unique on TenantId, system-default partial unique, FK CASCADE)
- 015: Invalid (TS and C# both use tamma_sk_; only CLAUDE.md prose mismatches — docs-only)
- 016: Fixed partial (api_keys: scope CHECK, active partial index, rotated_from self-FK; permissions text[]→jsonb deferred)
- 017: Fixed partial (github_installations: account_type CHECK, AppId int→bigint, AccountLogin/TenantId indexes)
- 018: Fixed (github_installation_repos: Owner/Name/CreatedAt/UpdatedAt columns + RepoFullName index + backfill)
- 019: Fixed partial (user_invites role CHECK; InviteUser handler rejects Guid.Empty fallback)
- 023: Invalid (intentional design divergence — EnsurePersonalTenantMiddleware replaces sentinel)
- 024: Fixed partial (users: GitHubId int→bigint, Settings jsonb, role+auth_method CHECK, LOWER(email) partial unique; Email kept NOT NULL)
- 025: Fixed (tenants: plan CHECK, deleted_at partial index)
- 026: Fixed partial (domain_events: TenantId+IssueNumber partial compound; RLS deferred to Phase-2)
- 027: Fixed (tenant_memberships role CHECK)
- 028: Invalid (intentional one-table-with-discriminator collapse)
- 029: Fixed partial (workflow_instances: TenantId+DefinitionId, TenantId+Status compound; RLS deferred)
- 030: Fixed partial (prompt_overrides: max_tokens+version CHECKs, Version, CreatedBy/UpdatedBy)
- 031: Fixed (agent_configs: split partial unique NULL-vs-NOT-NULL, FK CASCADE)
- 032: Fixed partial (provider_diagnostics: 8 columns + 5 indexes + budget partial; FK deferred per firehose model)
- 033: Fixed (provider_health: status CHECK, open partial, system-default partial unique)

FKs on api_keys/github_installations/user_invites.invited_by/domain_events/
workflow_instances are intentionally NOT added in this phase — the C# port's
webhook firehose + personal-tenant middleware produces orphan rows in normal
operation. RLS in Phase-2 (findings 020-022) will provide the equivalent
isolation guarantee.

Tests: 540 Tamma.Api.Tests + 882 Tamma.Activities.Tests + 7 Tamma.Core.Tests
all pass. Test fixtures updated in BudgetServiceTests, DiagnosticsAggregationTests,
AgentEndpointsIntegrationTests to seed tenant rows where new FKs require it.

+++ apps/tamma-elsa/src/Tamma.Data/Entities/GitHubInstallation.cs
@@ -5,8 +5,20 @@ public class GitHubInstallation
     public Guid Id { get; set; }
     public long InstallationId { get; set; }
     public string AccountLogin { get; set; } = null!;
+
+    /// <summary>
+    /// GitHub account type. Constrained to <c>User | Organization</c> via a
+    /// CHECK constraint (matches TS migration 001).
+    /// </summary>
     public string AccountType { get; set; } = null!;
-    public int AppId { get; set; }
+
+    /// <summary>
+    /// GitHub App ID. Widened to <c>bigint</c> (long) to match TS — App IDs
+    /// are bigint in the GitHub API, and narrowing to int32 leaves no
+    /// headroom.
+    /// </summary>
+    public long AppId { get; set; }
+
     public string? AppSlug { get; set; }
     public string Permissions { get; set; } = "{}";
     public DateTime? SuspendedAt { get; set; }
