+++ apps/tamma-elsa/src/Tamma.Data/Repositories/IUserRepository.cs
@@ -7,7 +7,7 @@ public interface IUserRepository
     Task<User> CreateAsync(User user);
     Task<User?> GetByIdAsync(Guid id);
     Task<User?> GetByEmailAsync(string email);
-    Task<User?> GetByGitHubIdAsync(int githubId);
+    Task<User?> GetByGitHubIdAsync(long githubId);
     Task<(List<User> Users, int Total)> ListAsync(int limit, int offset, string? role);
     Task<User> UpdateAsync(User user);
     Task SoftDeleteAsync(Guid id);
